update : paufinement

+++ srcs/binary_reader.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   binary_reader.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: awallet <marvin@42.fr>                     +#+  +:+       +#+        */
+/*   By: awallet <awallet@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/29 16:17:25 by awallet           #+#    #+#             */
-/*   Updated: 2022/05/29 18:07:20 by awallet          ###   ########.fr       */
+/*   Updated: 2022/05/30 15:42:52 by awallet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,14 +19,14 @@ unsigned int	ft_rbool(struct s_chaineoctet *self)
 	return (self->buffer[self->pos++]);
 }
 
-int	ft_rbyte(struct s_chaineoctet *self)
+signed int	ft_rbyte(struct s_chaineoctet *self)
 {
 	if (self->pos + 1 > self->len)
 		ft_bufferoverflow();
 	return (self->buffer[self->pos++] & 0xFF);
 }
 
-int	ft_rshort(struct s_chaineoctet *self)
+signed short	ft_rshort(struct s_chaineoctet *self)
 {
 	if (self->pos + 2 > self->len)
 		ft_bufferoverflow();
@@ -34,7 +34,7 @@ int	ft_rshort(struct s_chaineoctet *self)
 		| (self->buffer[self->pos++] & 0xFF));
 }
 
-int	ft_rint(struct s_chaineoctet *self)
+signed int	ft_rint(struct s_chaineoctet *self)
 {
 	if (self->pos + 4 > self->len)
 		ft_bufferoverflow();
@@ -44,7 +44,7 @@ int	ft_rint(struct s_chaineoctet *self)
 		| (self->buffer[self->pos++] & 0xFF));
 }
 
-int	ft_rlong(struct s_chaineoctet *self)
+signed long int	ft_rlong(struct s_chaineoctet *self)
 {
 	if (self->pos + 8 > self->len)
 		ft_bufferoverflow();
